Seven Samurai
<[Seven Samurai](https://letterboxd.com/film/seven-samurai/) - 

Fed up with the annual hit-and-run raids by a marauding gang of ferocious bandits, the peaceful farmers of a small sixteenth-century Japanese community summon up the courage to take matters into their own hands, dispatching the hot-headed villager, Rikichi, to bring help. When the farmer returns to the village, seven battle-scarred ronin become the defenceless hamlet's new protectors, entrusted with the challenging task of teaching the inexperienced peasants how to fight, in exchange for a handful of rice and lodging. Undoubtedly, before the imminent attack of the mounted plunderers, the odds are stacked up against the farmers and their defenders; however, both have sworn to fight like lions for their freedom. Can the Seven Samurai pull off a surprise victory?
—Nick Riganas